#2319 moved functions to core reader and fixed tab opening with navigation in custom
